Test error: 404 on TableListing ExposeConf, expecting a 401
